rework A/V
HDMI support features are moved to their native site, get rid of UsageConfig.
AudioSelection has been reworked, now you can control with the color buttons if you want a Dolby Digital Passthrough or Downwix, as well as having it configurable, plus other proc features are controlled.
It is added to be able to control the Dolby digital with a long press with the AUDIO key.
CanProc detection system added.

add Change track AC3 to Hotkey