UsageConfig.py: write delete line "registry.arm.bin" (if isfile in home)
@persianpros delete line in file crontab in the build environment...here we manage if the file "registry.arm.bin" exists then execute the actions for this file.
if "registry.arm.bin" exists and no such line exists, it will create it in the last crontab line and execute the remove action on reboots.
With this we ensure if the user deletes it, then create it and it will create it as long as (registry.arm.bin exists), meanwhile it will not.